Fix Call ID is not reflecting settings/vconf state.
When user is asking to hide his phone number to called
party, ID is not hidden even if network allows it.
This fix is temporary solution for Tizen_2.1 based on vconf
key.
In Tizen_3.0 release, complete solution will be:
- settings-app: on CLIR status changed, call new TAPI API.
- libslp-tapi: implement new API tel_set_ss_cli_status()
- tel-plugin-dbus_tapi: Add new DBus API in interface
- tel-plugin-imc: complete entry point .cli_activate() to
send AT+CLIR=1 / .cli_deactivate() to send AT+CLIR=0 to
modem.

Change-Id: I6da2f38e339b33a54fa1addb3f2dd49bffeb1b65